Implementacion ClientFutureReservation
Esta funcion ha sido añadida para impedir el borrado de una habitacion que contiene reservas posteriores al dia actual

Tambien se ha añadido un manejador de errores personalizado en caso de existir reservas futuras activas

+++ stayhub/src/main/java/com/Angelh0/stayhub/service/impl/RoomServiceImpl.java
@@ -1,23 +1,24 @@
 package com.Angelh0.stayhub.service.impl;
 
 import com.Angelh0.stayhub.converter.RoomConverter;
-import com.Angelh0.stayhub.dto.room.ResponseRoomDTO;
-import com.Angelh0.stayhub.dto.room.RoomAdminDTO;
-import com.Angelh0.stayhub.dto.room.RoomDTO;
-import com.Angelh0.stayhub.dto.room.UpdateRoomDTO;
+import com.Angelh0.stayhub.dto.room.*;
 import com.Angelh0.stayhub.entity.RoomEntity;
 import com.Angelh0.stayhub.enums.StatusType;
 import com.Angelh0.stayhub.exception.NotFoundException;
+import com.Angelh0.stayhub.exception.RoomException.RoomContainsReservation;
+import com.Angelh0.stayhub.grpcClient.GrpcClientFutureReservation;
 import com.Angelh0.stayhub.repository.AccommodationRepository;
 import com.Angelh0.stayhub.repository.RoomRepository;
 import com.Angelh0.stayhub.service.BusinessService;
 import com.Angelh0.stayhub.service.RoomService;
+import com.checkAvailability.grpc.ReservationResponse;
 import jakarta.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import com.Angelh0.stayhub.entity.AccommodationEntity;
 
 
+import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -38,6 +39,8 @@ public class RoomServiceImpl implements RoomService {
     @Autowired
     private BusinessService businessService;
 
+    @Autowired
+    private GrpcClientFutureReservation grpcClientFutureReservation;
 
 
     @Override
@@ -118,6 +121,10 @@ public class RoomServiceImpl implements RoomService {
         Optional<RoomEntity> room = roomRepository.findByUuid(uuid);
 
         if (room.isPresent()) {
+            RoomAvailabilityDTO availabilityDTO = grpcClientFutureReservation.getFutureReservation(room.get().getUuid());
+            if (availabilityDTO.isAvailable()) {
+                throw new RoomContainsReservation("La habitacion seleccionada, contiene reservas activas posteriores a la fecha actual." + LocalDate.now());
+            }
             roomRepository.deleteByUuid(uuid);
         }
     }
